refactor: moved getUser...Attendance to new usersToEvents.ts, temp schema optionals for sanity
