+++ packages/backend/convex/_generated/api.d.ts
@@ -15,6 +15,7 @@ import type * as data_ml_friendRecs from "../data_ml/friendRecs.js";
 import type * as data_ml_friends from "../data_ml/friends.js";
 import type * as data_ml_universal from "../data_ml/universal.js";
 import type * as data_ml_users from "../data_ml/users.js";
+import type * as data_ml_usersToEvents from "../data_ml/usersToEvents.js";
 import type * as eventsIngest from "../eventsIngest.js";
 import type * as seed from "../seed.js";
 import type * as users from "../users.js";
@@ -33,6 +34,7 @@ declare const fullApi: ApiFromModules<{
   "data_ml/friends": typeof data_ml_friends;
   "data_ml/universal": typeof data_ml_universal;
   "data_ml/users": typeof data_ml_users;
+  "data_ml/usersToEvents": typeof data_ml_usersToEvents;
   eventsIngest: typeof eventsIngest;
   seed: typeof seed;
   users: typeof users;
